pyproject.toml
[project]
name = "email_triage_env"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
  "openenv-core>=0.2.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.27.0",
  "pydantic>=2.6.0",
  "openai>=1.30.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
server = "server.app:main"

